Organize commands into separate scopes
In this case just have one for comments.
