+++ fioctl/utils.py
@@ -1,0 +1,30 @@
+from datetime import datetime
+
+def merge_streams(stream, other_stream, comparison=lambda x, y: x["id"] <= y["id"]):
+    fetch = lambda stream: next(stream, None)
+    head1, head2 = fetch(stream), fetch(other_stream)
+
+    while head1 and head2:
+        if comparison(head1, head2):
+            yield head1
+            head1 = fetch(stream)
+        else:
+            yield head2
+            head2 = fetch(other_stream)
+    
+    if head1:
+        yield head1
+        for head1 in stream:
+            yield head1
+    
+    if head2:
+        yield head2
+        for head2 in other_stream:
+            yield head2
+
+
+def datetime_compare(first, second):
+    return from_iso(first) <= from_iso(second)
+
+def from_iso(date_string):
+    return datetime.strptime(date_string, "%Y-%m-%dT%H:%M:%S.%fZ")
